requirements.txt
streamlit
yfinance
pandas
mplfinance
google-generativeai
python-dotenv
psycopg2-binary